changed file contorllers
