+++ .gitignore
@@ -5,6 +5,7 @@ node_modules
 .env.test.local
 .env.production.local
 .env
+.dockerfile
